feat: a Machine has a name
The commit adds a new feature to the Machine class: the ability to define a name for the machine. This is done by updating the define method in Machine.php to accept an optional $definition parameter that includes the name field. The State class is also updated to include a public property $name that is set when a new State instance is created. Finally, a new test is added in MachineTest.php to ensure that a Machine instance has a name when defined with a name.