fix: Set a minimum version number on State initialization and add negative version test
The State class now checks if the version number is less than 1, if so, it sets it to the minimum value of 1. The test in UselessMachinesTest.php checks whether the Machine class handles a negative version number correctly, by initializing a machine with a negative version number, and checking that the version number is correctly set to the minimum value of 1.