feat: Allow description field in machine definition
This commit adds support for a new "description" field in the machine definition that can be used to describe the purpose of the machine. If the description is empty, it will be set to null.

Also, the constructor for the State class now accepts a new "description" parameter. If the description is not provided, it will be set to null.