feat: Add events list to EventMachine and StateNode
Add a new public property $events to EventMachine and StateNode classes. The property contains a list of all the event types accepted by the machine and its descendants.

Also, add a new test case to MachineTest.php to test the new feature.

This is a feature addition as it adds new functionality to the existing codebase.