refactor: simplify state node type determination
Simplify the logic for determining the state node type by using the null coalescing operator instead of the match expression.

The _initialize() method that was removed was empty and didn't serve any purpose.

The getEvents() method was not modified in this patch, but a TODO comment was added to consider caching events in the future.