feat: add support for machine names in definition
The MachineDefinition class now supports a name property to be set in the definition array. If a name is not provided, the default name is used. The define() method was updated to handle the new parameter.

Add a new test that checks if a machine has a name when defined.

Changes include:

Addition of name property in MachineDefinition class
Modification of define() method in MachineDefinition class to handle the new parameter
Addition of a new test in MachineTest.php to check if a machine has a name