test: Update MachineTest to add more test cases
Add more test cases to the MachineTest to ensure proper behavior of the MachineDefinition class.

The commit adds the following changes:

Check that an instance of MachineDefinition is created when no parameters are passed.
Check that a machine with a given name can be created.
Check that a machine with a null name has the default name.