refactor: rename $definition to $config in MachineDefinition::define method
This commit renames the $definition parameter to $config in the MachineDefinition::define method to make it more descriptive. It also updates the corresponding tests to use the new parameter name.

The $config parameter now represents an optional configuration array that can be passed to the define method to set the machine's properties such as its name.