feat: Add machine config property to MachineDefinition
This commit adds a new config property to the MachineDefinition class, which can be set by passing an array to the define() method. The config property is then stored in the $machineWithName->config property, which is verified by a new test case. If no config array is provided, the config property is set to null.