feat: Add version property to MachineDefinition
This commit adds a new version property to the MachineDefinition class, which can be set by passing a version string to the define() method. The version property is then stored in the $machineWithVersion->version property, which is verified by a new test case. If no version string is provided, the version property is not set and is null.