test: Refactor MachineDefinition null config test
This commit refactors the test case for the version property of a MachineDefinition object to use a null configuration, by creating a new MachineDefinition object without passing any arguments to the define() method. The new nullMachine object is then used in the expectations, which use the toHaveProperty() assertion to check that the version property is present and the toBeNull() assertion to check that its value is null. This change simplifies the test case and makes it easier to read.