feat: add root state definition to machine definition
Add a root state definition to the MachineDefinition class to allow users to define the initial state of their state machine. This is achieved by creating a new StateDefinition instance and assigning it to the root property of the MachineDefinition instance.

This commit also adds a new StateDefinition class, which is used to define the structure of a state within a state machine.