feat(StateDefinition): add globalId property to StateDefinition class
Added a new public property named 'globalId' to the StateDefinition class. This property holds the unique global identifier of the state definition. The global identifier is generated by either using the 'id' value from the config or generating a unique ID by merging the machine ID with the path, separated by the machine delimiter. Additionally, a new unit test has been added to test the globalId property of the root state definition.