feat(StateDefinition): add description property to StateDefinition class
Added a new public property named 'description' to the StateDefinition class. This property holds the description of the state definition. The description can be passed as a configuration option when defining a new state definition. Additionally, a new unit test has been added to test the description property of the root state definition.