refactor: use SplObjectStorage for MachineDefinition's idMap
This commit refactors the idMap property of the MachineDefinition class to use a SplObjectStorage object instead of an associative array. This is to improve performance and readability, as SplObjectStorage is designed specifically for storing objects as keys. The commit also updates the corresponding test in MachineDefinitionTest to ensure that the idMap is an instance of SplObjectStorage, and that it contains the root state definition with a count of 1. Finally, the StateDefinition class is updated to use the attach() method of SplObjectStorage instead of setting the key directly in the idMap property.