refactor: Update StateDefinition properties
The 'localId' property of the StateDefinition class has been updated to 'key' to better reflect its purpose of representing the location of the state definition in the overall state value. This update was made in the MachineDefinition, StateDefinition, and corresponding test files.

Note: This commit may have potential breaking changes for any code that relied on the previous 'localId' property.