feat: add initial state to MachineDefinition
This commit adds an initialState property to MachineDefinition class, which is used to hold the initial state definition for the state machine. This property is populated by setting the value of initialState property of the root StateDefinition. Additionally, a findInitialState() method is added to StateDefinition class to help in the initialization of initialState property.

A new test InitialStateTest is also added to test the functionality of the added initialState property.