feat: add support for entry actions in state definitions
This commit adds support for entry actions in state definitions. The StateDefinition class now has an entry property that can be used to specify one or more actions to be executed upon entering the state definition. Entry actions can be specified in the entry key of the state definition configuration array.

The commit also includes a new test in StateDefinitionTest.php that checks that states can have entry actions.