feat: add support for exit actions in state definitions
This commit adds support for exit actions in state definitions. The StateDefinition class now has an exit property that can be used to specify one or more actions to be executed upon exiting the state definition. Exit actions can be specified in the exit key of the state definition configuration array.

The commit also includes a new test in StateDefinitionTest.php that checks that states can have exit actions.