test: ability to remove a key from context data in ContextDefinitionTest
This commit adds a new test case to the ContextDefinitionTest file to test the ability to remove a key from context data using the remove method of the ContextDefinition class. The test case creates a new context definition, sets two key-value pairs, removes one of the keys using the remove method, and checks that the key has been removed and the other key still exists in the context data.