refactor: use null coalescing operator for target property initialization
This commit refactors the TransitionDefinition class to use the null coalescing operator (??) instead of the isset and strict comparison (===) operators when initializing the target property. It assigns the value of $this->transitionConfig['target'] to $targetConfig with the null coalescing operator and then checks if $targetConfig is null. If it is null, the target property is set to null. Otherwise, it is set to $this->source->parent->states[$targetConfig].