feat: Add support for entry actions in state definitions
This commit adds support for running entry actions in state definitions. It introduces a new method runEntryActions() to StateDefinition class which is called when transitioning to a new state. It also adds a new method runAction() to MachineDefinition class which can be used to run actions defined in the machine's behavior.

The new feature is also covered by a new test in EntryActionsTest.php.

The commit also includes some formatting changes to the code, but no functional changes are made.