feat(machine): Add support for exit actions in state definitions
Add a new runExitActions() method to the StateDefinition class, which executes all exit actions of the current state definition with the given event. The runExitActions() method is called from the MachineDefinition class when transitioning from one state to another.

Also, update the $transitionDefinition variable in the MachineDefinition class to be nullable, by updating its PHPDoc block.

Finally, add a new test case in the ExitActionsTest class to ensure that exit actions are executed when transitioning from a state.

This commit introduces new functionality that allows developers to execute exit actions when transitioning from a state, thereby enabling them to clean up or perform certain actions before leaving a state.