refactor: Improve runAction() and add getActionBehavior() methods
This commit refactors the runAction() method to be more clear and
understandable. It now retrieves the appropriate action behavior based
on the action definition using a new getActionBehavior() method.

The getActionBehavior() method checks if the action definition is an
invokable ActionBehavior subclass and creates a new instance if it
is. Otherwise, it looks up the action definition in the machine
behavior and retrieves the corresponding behavior. If the retrieved
behavior is not callable, a new instance is created.