refactor: remove unused EventMachine::transition() method
The EventMachine::transition() method is unused and has been removed.

The EventMachine class is an abstract base class for creating state machines. It provides the definition() method that subclasses should override to provide a specific MachineDefinition. The transition() method was previously provided by EventMachine, but it has been removed because it was unused.