feat: Add event type behavior to TrafficLightsMachine
Add support for event type behavior to the TrafficLightsMachine by defining event classes that extend the EventDefinition class. The new initializeEvent() method in MachineDefinition is used to initialize events, and is called when transition() is called.

In addition to the new behavior, new event classes are added to the TrafficLightsMachine for the "MUT", "INC", and "DEC" event types. These event classes extend the EventDefinition class.