refactor: Rename definition() to build() in EventMachine and TrafficLightsMachine
This commit renames the definition() method to build() in the EventMachine and TrafficLightsMachine classes to better reflect their purpose. The build() method creates and returns a new MachineDefinition instance, which is used to define the state machine's configuration.

This change affects the tests/EventMachineTest.php and tests/Stubs/Machines/TrafficLights/TrafficLightsMachine.php files, where the method calls are updated to use the new name.