feat: Change ContextManager constructor to use optional type
The ContextManager constructor has been modified to accept an optional type with the help of Spatie\LaravelData\Optional. This change allows for better type hinting and makes it clear that the data parameter is optional. The tests have been updated to reflect this change.

Also, minor documentation update in src/Definition/MachineDefinition.php.