feat: set default type if not provided in EventBehavior
This commit adds functionality to the EventBehavior class to set the default type if it is not provided. If type is null, it will be set to the static getType() method of the class. This is useful when creating new instances of EventBehavior and not having to set the type manually.