feat(TrafficLightsMachine): Add SubtactValueEvent and SubtractValueAction
This commit adds a new SubtactValueEvent to TrafficLightsMachine and registers it in the events array of the machine. Also, the SubtractValueAction is added to the actions array of the machine. When the SubtactValueEvent is triggered, it calls the SubtractValueAction to subtract the specified value from the count property of the TrafficLightsContext.